refactor(mcp): drain managed-document (doc_*) tool domain into a scoped pack (EP-8DC217EB BET-4)
Move the seven managed-document MCP tools (doc_save, doc_load, doc_search,
doc_link, doc_version_list, doc_state_change, doc_list_references) out of the
16k-line mcp-tools.ts executeTool switch into a scoped ToolPack at
apps/web/lib/mcp/packs/document-pack.ts, following the existing pack pattern.

- Definitions moved verbatim out of PLATFORM_TOOLS; handlers reproduce the
  former switch cases exactly (same coercion, DB calls, and result envelopes).
- The doc-only helper resolveDocumentActorPrincipalId moved into the pack;
  the shared optionalString/stringArray coercers are copied locally so the pack
  owns its inputs without depending on mcp-tools.ts internals.
- grants mirror agent-grants.ts TOOL_TO_GRANTS (still the gating source).
- Registered documentPack in the composeToolPacks registry.
- Shrank the tool-pack inline-case baseline by the seven drained names and
  retightened the module-size baseline (mcp-tools.ts 16554 -> 16260 LOC).
- Added document-pack.test.ts asserting tool names, grant mirror, provenance-free
  descriptions, and per-handler envelope shapes.

Behavior-preserving relocation: the drained tools dispatch through the pack
registry (after the same kernel gate) and behave identically when called via MCP.

Process-Spine-Decision: EP-8DC217EB plan §4 BET-4 — drain the doc tool domain from the executeTool switch into a scoped pack; ratchet baseline shrunk; behavior-preserving relocation.
