+++ scripts/mcp-tool-pack-baseline.json
@@ -52,13 +52,6 @@
   "diagnose_sandbox",
   "discovery_sweep",
   "dismiss_entity",
-  "doc_link",
-  "doc_list_references",
-  "doc_load",
-  "doc_save",
-  "doc_search",
-  "doc_state_change",
-  "doc_version_list",
   "dpf_test_kernel_refuse_probe",
   "draft_marketing_asset",
   "drive_browser_task",
